Поддержка функции добавления одинакового товара в заказ как разные товарные позиции из CRM

+++ intaro.retailcrm/classes/general/history/RetailCrmHistory_v5.php
@@ -126,7 +126,7 @@ class RetailCrmHistory
                     }
 
                     if ($registerNewUser === true) {
-                        $userPassword = uniqid("R");
+                        $userPassword = uniqid();
 
                         $arFields = array(
                             "EMAIL"             => $customer['email'],
@@ -137,12 +137,12 @@ class RetailCrmHistory
                         );
                         $registeredUserID = $newUser->Add($arFields);
                         if ($registeredUserID === false) {
-                            RCrmActions::eventLog('RetailCrmHistory::orderHistory', 'CUser::Register', 'Error register user: ' . $newUser->LAST_ERROR);
+                            RCrmActions::eventLog('RetailCrmHistory::orderHistory', 'CUser::Register', 'Error register user');
                             continue;
                         }
 
                         if(RCrmActions::apiMethod($api, 'customersFixExternalIds', __METHOD__, array(array('id' => $customer['id'], 'externalId' => $registeredUserID))) == false) {
-                             continue;
+                            continue;
                         }
                     }
 
@@ -361,7 +361,7 @@ class RetailCrmHistory
                         }
 
                         if ($registerNewUser === true) {
-                            $userPassword = uniqid("R");
+                            $userPassword = uniqid();
 
                             $newUser = new CUser;
                             $arFields = array(
@@ -384,7 +384,7 @@ class RetailCrmHistory
                             $registeredUserID = $newUser->Add($arFields);
 
                             if ($registeredUserID === false) {
-                                RCrmActions::eventLog('RetailCrmHistory::orderHistory', 'CUser::Register', 'Error register user ' . $newUser->LAST_ERROR);
+                                RCrmActions::eventLog('RetailCrmHistory::orderHistory', 'CUser::Register', 'Error register user');
 
                                 continue;
                             }
@@ -423,7 +423,7 @@ class RetailCrmHistory
                 if (isset($order['externalId'])) {
                     $itemUpdate = false;
 
-                    if ($order['externalId'] && is_numeric($order['externalId'])) {
+                    if ($order['externalId']) {
                         try {
                             $newOrder = Bitrix\Sale\Order::load($order['externalId']);
                         } catch (Bitrix\Main\ArgumentNullException $e) {
@@ -566,9 +566,6 @@ class RetailCrmHistory
                                 }
                             } elseif (array_key_exists($key, $order['delivery']['address'])) {
                                 if ($propsKey[$orderProp]['TYPE'] == 'LOCATION') {
-                                    if( $order['delivery']['address']['index'] ) {
-                                        $location = CSaleLocation::GetByZIP($order['delivery']['address']['index']);
-                                    }
                                     $order['delivery']['address'][$key] = trim($order['delivery']['address'][$key]);
                                     if(!empty($order['delivery']['address'][$key])){
                                         $parameters = array();
@@ -585,9 +582,7 @@ class RetailCrmHistory
                                         $parameters['filter']['NAME.LANGUAGE_ID'] = 'ru';
 
                                         try {
-                                            if ( !isset($location) ) {
-                                                $location = \Bitrix\Sale\Location\Search\Finder::find($parameters, array('USE_INDEX' => false, 'USE_ORM' => false))->fetch();
-                                            }
+                                            $location = \Bitrix\Sale\Location\Search\Finder::find($parameters, array('USE_INDEX' => false, 'USE_ORM' => false))->fetch();
                                             $somePropValue = $propertyCollection->getItemByOrderPropertyId($propsKey[$orderProp]['ID']);
                                             self::setProp($somePropValue, $location['CODE']);
                                         } catch (\Bitrix\Main\ArgumentException $argumentException) {
@@ -611,7 +606,7 @@ class RetailCrmHistory
                         foreach ($optionsLegalDetails[$personType] as $key => $orderProp) {
                             if (array_key_exists($key, $order)) {
                                 $somePropValue = $propertyCollection->getItemByOrderPropertyId($propsKey[$orderProp]['ID']);
-                                self::setProp($somePropValue, RCrmActions::fromJSON($order[$key]));
+                                self::setProp($somePropValue, $order[$key]);
                             }
                         }
                     }
@@ -646,11 +641,32 @@ class RetailCrmHistory
                     if (isset($order['items'])) {
                         $itemUpdate = true;
 
+                        $response = RCrmActions::apiMethod($api, 'orderGet', __METHOD__, $order['id']);
+                        if (isset($response['order'])) {
+                            $orderTemp = $response['order'];
+                            $ditems = [];
+                            foreach ($orderTemp['items'] as $item) {
+                                $ditems[$item['offer']['xmlId']]['quantity'] += $item['quantity'];
+                                $ditems[$item['offer']['xmlId']]['discountTotal'] += $item['quantity'] * $item['discountTotal'];
+                                $ditems[$item['offer']['xmlId']]['initialPrice'] = (float)$item['initialPrice'];
+                                $ditems[$item['offer']['xmlId']]['price_sum'] = $ditems[$item['offer']['xmlId']]['initialPrice'] * $ditems[$item['offer']['xmlId']]['quantity'] - $ditems[$item['offer']['xmlId']]['discountTotal'];
+                                $ditems[$item['offer']['xmlId']]['price_item'] = $ditems[$item['offer']['xmlId']]['price_sum'] / $ditems[$item['offer']['xmlId']]['quantity'];
+                            }
+                            unset($orderTemp);
+                        }
+
+                        $log->write($ditems, 'duplicateItemsOrderHistory');
+
                         foreach ($order['items'] as $product) {
+                            if($ditems[$product['offer']['xmlId']]['quantity']){
+                                $product['quantity'] = $ditems[$product['offer']['xmlId']]['quantity'];
+                            }
+
                             $item = self::getExistsItem($basket, 'catalog', $product['offer']['externalId']);
 
                             if (!$item) {
                                 if ($product['delete']) {
+
                                     continue;
                                 }
 
@@ -679,9 +695,11 @@ class RetailCrmHistory
                             }
 
                             if ($product['delete']) {
-                                $item->delete();
+                                if ($ditems[$product['offer']['xmlId']]['quantity'] <= 0) {
+                                    $item->delete();
 
-                                continue;
+                                    continue;
+                                }
                             }
 
                             if ($product['quantity']) {
@@ -709,6 +727,12 @@ class RetailCrmHistory
                                     $item->setField('DISCOUNT_VALUE', '');
                                     $item->setField('DISCOUNT_PRICE', $resultDiscount);
                                     $item->setField('PRICE', $itemCost - $resultDiscount);
+
+                                    //set price dublicate item
+                                    if ($ditems[$product['offer']['xmlId']]['price_item']) {
+                                        $item->setField('PRICE', $ditems[$product['offer']['xmlId']]['price_item']);
+                                        $item->setField('DISCOUNT_PRICE', '');
+                                    }
                                 }
                             }
                         }
@@ -772,16 +796,20 @@ class RetailCrmHistory
                     if (!empty($newHistoryPayments)) {
                         foreach ($newOrder->getPaymentCollection() as $orderPayment) {
                             if (array_key_exists($orderPayment->getField('XML_ID'), $newHistoryPayments)) {
+
                                 $paymentId = $orderPayment->getId();
                                 $paymentExternalId = RCrmActions::generatePaymentExternalId($paymentId);
                                 if (is_null($paymentId)) {
                                     RCrmActions::eventLog('RetailCrmHistory::orderHistory', 'paymentsUpdate', 'Save payment error, order=' . $order['number']);
                                     continue;
                                 }
+
+                                $paymentExternalId = $orderPayment->getId();
+
                                 if ($paymentExternalId) {
                                     $newHistoryPayments[$orderPayment->getField('XML_ID')]['externalId'] = $paymentExternalId;
                                     RCrmActions::apiMethod($api, 'paymentEditById', __METHOD__, $newHistoryPayments[$orderPayment->getField('XML_ID')]);
-                                    \Bitrix\Sale\Internals\PaymentTable::update($paymentId, array('XML_ID' => ''));
+                                    \Bitrix\Sale\Internals\PaymentTable::update($paymentExternalId, array('XML_ID' => ''));
                                 }
                             }
                         }
@@ -813,6 +841,26 @@ class RetailCrmHistory
             //new filter
             $historyFilter['sinceId'] = $end['id'];
         }
+    }
+
+    /**
+     * @param $array
+     * @param $value
+     *
+     * @return array
+     */
+    public static function search_array_by_value($array, $value)
+    {
+        $results = array();
+        if (is_array($array)) {
+            $found = array_search($value,$array);
+            if ($found) {
+                $results[] = $found;
+            }
+            foreach ($array as $subarray)
+                $results = array_merge($results, static::search_array_by_value($subarray, $value));
+        }
+        return $results;
     }
 
     public static function assemblyCustomer($customerHistory)
@@ -1017,9 +1065,10 @@ class RetailCrmHistory
                         $serviceCode = str_replace(array('-'), "_", $serviceCode);
                     }
                 }
+
                 if ($deliveryCode) {
                     try {
-                        $deliveryService = \Bitrix\Sale\Delivery\Services\Manager::getObjectByCode($deliveryCode . ':' . $serviceCode);
+                        $deliveryService = \Bitrix\Sale\Delivery\Services\Manager::getObjectByCode($deliveryCode . ':' . $orderCrm['delivery']['service']['code']);
                     } catch (Bitrix\Main\SystemException $systemException) {
                         RCrmActions::eventLog('RetailCrmHistory::deliveryEdit', '\Bitrix\Sale\Delivery\Services\Manager::getObjectByCode', $systemException->getMessage());
                     }
@@ -1036,13 +1085,13 @@ class RetailCrmHistory
 
         if ($delivery) {
             if (!$update) {
-                    $shipment = $shipmentColl->createItem($delivery);
-                    $shipment->setFields(array(
-                        'BASE_PRICE_DELIVERY' => $orderCrm['delivery']['cost'],
-                        'CURRENCY' => $order->getCurrency(),
-                        'DELIVERY_NAME' => $delivery->getName(),
-                        'CUSTOM_PRICE_DELIVERY' => 'Y'
-                    ));
+                $shipment = $shipmentColl->createItem($delivery);
+                $shipment->setFields(array(
+                    'BASE_PRICE_DELIVERY' => $orderCrm['delivery']['cost'],
+                    'CURRENCY' => $order->getCurrency(),
+                    'DELIVERY_NAME' => $delivery->getName(),
+                    'CUSTOM_PRICE_DELIVERY' => 'Y'
+                ));
             } else {
                 foreach ($shipmentColl as $shipment) {
                     if (!$shipment->isSystem()) {
@@ -1157,8 +1206,7 @@ class RetailCrmHistory
         foreach ($paymentsCrm['payments'] as $paymentCrm) {
             if (isset($paymentCrm['externalId']) && !empty($paymentCrm['externalId'])) {
                 //find the payment
-                $nowPaymentId = RCrmActions::getFromPaymentExternalId($paymentCrm['externalId']);
-                $nowPayment = $paymentsList[$nowPaymentId];
+                $nowPayment = $paymentsList[$paymentCrm['externalId']];
                 //update data
                 if ($nowPayment instanceof \Bitrix\Sale\Payment) {
                     $nowPayment->setField('SUM', $paymentCrm['amount']);
@@ -1170,7 +1218,7 @@ class RetailCrmHistory
                         $nowPayment->setField('PAID', $optionsPayment[$paymentCrm['status']]);
                     }
 
-                    unset($paymentsList[$nowPaymentId]);
+                    unset($paymentsList[$paymentCrm['externalId']]);
                 }
             } elseif (array_key_exists($paymentCrm['type'], $optionsPayTypes)) {
                 $newHistoryPayments[$paymentCrm['id']] = $paymentCrm;
